+++ examples/basic_demo.py
@@ -13,12 +13,12 @@ from delineator.util import write_outputs
 
 
 def tryme():
-    config = DelineatorConfig(verbose=True, output_format="geojson", fill=False, rivers=False, outlets=False, snapping=False)
+    config = DelineatorConfig(verbose=True, fill=False, rivers=False, outlets=False, snapping=False)
     #w, r, o = delineate(44.705, 4.206, config)  # France (23,
     #w, r, o = delineate(63.938, -21.004, config)  # Iceland (27, built in)
     #w, r, o = delineate(-0.854, 41.350, config)  # East Africa (11)
     #w, r, o = delineate(48.982, 7.011, config)
-    w, r, o = delineate(64.69333333, -21.41083333, config)
+    w, r, o = delineate(65.85166667, -17.8975, config)
 
     write_outputs(w, r, o, config)